feat(refacto): moved ecsFargate.go into fargateService.go
