+++ scenarios/aws/fakeintake.go
@@ -3,11 +3,11 @@ package aws
 import (
 	ddfakeintake "github.com/DataDog/test-infra-definitions/components/datadog/fakeintake"
 	resourcesAws "github.com/DataDog/test-infra-definitions/resources/aws"
-	"github.com/DataDog/test-infra-definitions/resources/aws/ecs"
+	ecsFargate "github.com/DataDog/test-infra-definitions/resources/aws/ecs"
 )
 
 func NewEcsFakeintake(env resourcesAws.Environment) (*ddfakeintake.ConnectionExporter, error) {
-	fakeintake, err := ddfakeintake.NewECSFargateInstance(env, ecs.OldFargate{})
+	fakeintake, err := ecsFargate.NewECSFargateInstance(env)
 	if err != nil {
 		return nil, err
 	}
